Worked around stupid HSQLDB semantics when accessing in-memory databases.
HSQLDB tracks all databases (even in memory ones) by alias and will
return the original handle upon trying to make a new connection to the
same alias.  It appears there's no way to disable this behaviour with in
memory databases so we use semi-unique aliases on creation of memory
databases to avoid this issue.

Also fixed an issue where SHUTDOWN was not being called on databases
which would actually leak in this alias lookup table.