Major changes to dist packaging and introduced swt.jar.
We no longer try to bundle all dependencies into a single jar file for
execution with java -jar.  Instead, we take a more pragmatic approach
which distributes all libraries separately and includes an invocation
script to get the right java command-line constructed for the user.

These changes were ultimately necessary to support swt.jar as part of
the distribution package.